Lib and testing project for MidasGUI
error lnk 1102, somewhere in the project setting i think i made a
mistake. the lib is compiling.

+++ PMMain/Midas/MidasGUI/MyoTranslationFilter.h
@@ -106,7 +106,7 @@ private:
     */
     point getMouseUnitVelocity(float pitch, float yaw);
 
-    int keySelectAngle(orientation_data data);
+    float keySelectAngle(orientation_data data);
 
     ControlState* controlStateHandle;
     midasMode previousMode;
